+++ src/kgmk/dsa/number_theory/prime_factorize/with_prime_numbers/jit.py
@@ -32,7 +32,7 @@ def prime_factorize_factorial(
   pn: np.array,
 ) -> np.array:
   prime, cnt = [], []
-  idx = np.full(n + 1, -1, dtype=np.int32)
+  idx = np.full(n + 1, -1, dtype=np.int64)
   for i in range(n + 1):
     for p, c in prime_factorize(i, pn):
       i = idx[p]
